Created a single train script with both Arcface and Crossentropy loss
